Validation

+++ src/main/java/com/footballnewsmanager/backend/repositories/LeagueRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface LeagueRepository extends JpaRepository<League, Integer> {
+public interface LeagueRepository extends JpaRepository<League, Long> {
 
 
     Optional<League> findByApisportid(int id);
