+++ src/main/java/com/footballnewsmanager/backend/repositories/MarkerRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface MarkerRepository extends JpaRepository<Marker, Integer> {
+public interface MarkerRepository extends JpaRepository<Marker, Long> {
 
     Optional<Marker>findByName(String name);
