+++ src/main/java/com/footballnewsmanager/backend/repositories/SiteRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface SiteRepository extends JpaRepository<Site, Integer> {
+public interface SiteRepository extends JpaRepository<Site, Long> {
 
     Optional<Site> findByName(String name);
 }
